requirements.txt
Flask==2.2.5
gunicorn==20.1.0
gspread==5.12.0
google-auth==2.33.0
